Finishing up the test kit
